backend/requirements.txt
# VacaPlan AI — Backend Dependencies

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# LLM & Agent orchestration
anthropic==0.34.2
langgraph==0.2.28
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.1

# Vector store (open-source)
chromadb==0.5.11

# Data validation
pydantic==2.9.2

# HTTP client
httpx==0.27.2

# Payments
stripe==10.12.0

# Database
asyncpg==0.29.0
redis==5.1.1

# Google Calendar API
google-auth==2.35.0
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1

# PII detection (open-source, Microsoft)
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Utilities
python-dotenv==1.0.1
